feat: Add AlertControllerWrapper.swift for presenting alerts in Trilium Client
This commit adds a new file, AlertControllerWrapper.swift, which is responsible for presenting alerts in the Trilium Client app. It provides a SwiftUI wrapper for the UIAlertController class in UIKit. The wrapper allows for easy presentation of alerts with a title, message, text field, and save action. This will be used to display a dialog for entering a link URL in the StyleToolbar view.