refactor: Add NSAttributedString extension for converting to HTML
This commit adds a new file, NSAttributedString+Extension.swift, which provides an extension to the NSAttributedString class. The extension includes a method to convert an attributed string to HTML format. It removes XML declaration, DOCTYPE, html and body tags, and trims whitespace from the HTML string. This functionality will be used in the Trilium Client app for converting attributed strings to HTML before saving them to the database.